// src/shared/services/api/models/paginatedEnrollmentsDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Learning Assessment Portal API
 * OpenAPI spec version: v1
 */
import type { EnrollmentDetailDto } from './enrollmentDetailDto';

export interface PaginatedEnrollmentsDto {
  /** @nullable */
  data?: EnrollmentDetailDto[] | null;
  total?: number;
  page?: number;
  page_size?: number;
}